Added an initial version of a parser module for s-expressions. For our purposes, sexpr have to have an atom at first element of the list. The sexpr given as string to the Parse method are converted into lists of predicates.
Predicates consist of a name (a string) and a list of boost:any
elements. In practice, these elements are either strings or lists
containing elements (i.e. strings or lists containing elements
(i.e. ... you get it)). This kind of representation is suitable for
our effectors.

The Generate method of the parser converts lists of elements into an
sexpr string. Currently, the user of the Generate method has to take
care of properly double-quoting strings containing whitespace or
single quoting strings that look like s-expressions and so on.